+++ packages/react-scripts/package.json
@@ -52,6 +52,7 @@
     "postcss-loader": "1.3.3",
     "promise": "7.1.1",
     "react-dev-utils": "^0.5.2",
+    "stack-frame-overlay": "^0.4.0",
     "style-loader": "0.16.1",
     "url-loader": "0.5.8",
     "webpack": "2.4.1",
