+++ packages/stack-frame-overlay/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["react-app"]
+}
