+++ packages/stack-frame-overlay/.npmignore
@@ -1,0 +1,3 @@
+__tests__
+*.test.js
+*.spec.js
