+++ packages/stack-frame-overlay/README.md
@@ -1,0 +1,7 @@
+# `stack-frame-overlay`
+
+An overlay for displaying stack frames.
+
+# API
+
+<!-- Generated by documentation.js. Update this documentation by updating the source code. -->
