+++ packages/stack-frame-overlay/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "stack-frame-overlay",
+  "version": "0.4.0",
+  "description": "An overlay for displaying stack frames.",
+  "main": "lib/index.js",
+  "scripts": {
+    "prepublishOnly": "npm run build:prod && npm test",
+    "start": "NODE_ENV=development npm run build -- --watch",
+    "test": "exit 0",
+    "build": "babel src/ -d lib/",
+    "build:prod": "NODE_ENV=production babel src/ -d lib/"
+  },
+  "repository": "facebookincubator/create-react-app",
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/facebookincubator/create-react-app/issues"
+  },
+  "keywords": [
+    "overlay",
+    "syntax",
+    "error",
+    "red",
+    "box",
+    "redbox",
+    "crash",
+    "warning"
+  ],
+  "author": "Joe Haddad <timer150@gmail.com>",
+  "files": [
+    "lib/"
+  ],
+  "dependencies": {
+    "anser": "^1.2.5",
+    "babel-code-frame": "^6.22.0",
+    "stack-frame": "0.4.0",
+    "stack-frame-mapper": "0.4.0",
+    "stack-frame-parser": "0.4.0",
+    "stack-frame-unmapper": "0.4.0"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-eslint": "7.x",
+    "babel-preset-react-app": "^2.2.0",
+    "eslint": "^3.16.1",
+    "eslint-config-react-app": "^0.6.2",
+    "eslint-plugin-flowtype": "^2.21.0",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.4.1",
+    "react-dev-utils": "^0.5.2"
+  }
+}
