+++ packages/stack-frame-overlay/src/index.js
@@ -1,0 +1,9 @@
+/* @flow */
+import { inject, uninject } from './overlay';
+
+inject();
+if (module.hot && typeof module.hot.dispose === 'function') {
+  module.hot.dispose(function() {
+    uninject();
+  });
+}
