+++ packages/stack-frame-overlay/src/utils/dom/consumeEvent.js
@@ -1,0 +1,9 @@
+/* @flow */
+function consumeEvent(e: Event) {
+  e.preventDefault();
+  if (typeof e.target.blur === 'function') {
+    e.target.blur();
+  }
+}
+
+export { consumeEvent };
